pyproject.toml
# Create a pyproject.toml file for the project
# This file is used to define the project metadata and dependencies
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "fastapi-course"
version = "0.1.0"
description = "A FastAPI course project"
readme = "README.md"
requires-python = ">=3.10"

# Define the dependencies for the project
dependencies = [
    "pydantic==2.11.7",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "email_validator==2.2.0",
    "passlib==1.7.4",
    "bcrypt==4.3.0",
    "pyjwt==2.10.1",
    "python-multipart==0.0.20",
    "uvicorn==0.34.2",
    "httpx==0.28.1",
    "python-dotenv==1.1.0",
    "psycopg2-binary==2.9.10",
    "SQLAlchemy==2.0.41",
    "sqlmodel==0.0.24"
]

[project.optional-dependencies]
dev = [
    "ipython",
    "pytest",
    "black",
]